requirements.txt
requests
urllib3
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
jinja2

